refactorice el codigo separando en un componente aparte funcionalidades que no eran necesarias tenerlas juntas
