+++ package.json
@@ -12,6 +12,7 @@
     "@emailjs/browser": "^3.11.0",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
+    "@formkit/auto-animate": "^0.7.0",
     "@mui/icons-material": "^5.14.3",
     "@mui/material": "^5.14.3",
     "@tailwindcss/forms": "^0.5.3",
