+++ src/components/organisms/cartEmpty.js
@@ -4,7 +4,7 @@ import Image from 'next/image'
 
 const CartEmpty = () => {
   return (
-    <div className='flex flex-col items-center'>
+    <div className='flex flex-col items-center absolute right-0 bg-white text-black p-5 rounded-[10px]'>
         <Image width={250} src={carritoTriste}/>
         <h3 className='text-center font-bold'>Tu carrito esta vacio, <br/> ¡LLENALO DE VIAJES IMPERDIBLES!</h3>
     </div>
